+++ fixtures/basic/tests/snapshots.spec.ts
@@ -1,5 +1,0 @@
-import { expect, test } from '@playwright/test';
-
-test('snapshot test', async () => {
-  expect('hello playwright').toMatchSnapshot('greeting.txt');
-});